requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
sentence-transformers>=2.2.2
PyMuPDF>=1.23.0

faiss-cpu>=1.7.0
numpy>=1.26.4
urllib3>=2.0.0

groq>=0.4.1
python-dotenv>=1.0.0